+++ Kits/FirebaseKit/Sources/Domain/Entities/UserDetailEntity.swift
@@ -6,5 +6,5 @@ import Foundation
 public struct UserDetailEntity: Codable {
     public var favoriteExercise: String?
     public var totalTrainings: Int?
-    public var totalMinutesOfTraining: Double?
+    public var totalMinutesOfTraining: Int?
 }
